pyproject.toml
[project]
name = "aura"
version = "0.1.0"
description = "Personal Assistant on WhatsApp"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "requests",
    "APScheduler",
    "supabase",
    "python-dotenv",
    "elevenlabs",
    "openai",
    "pydantic",
    "realtime",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
